refactor: reduce eipfs and ipni stack to indexer stack
