+++ indexer/functions/handle-block-advert-publish-message.js
