+++ indexer/functions/handle-block-index-writer-message.js
