+++ indexer/lib/block-advert-publisher.js
