+++ indexer/lib/block-index-writer.js
