+++ indexer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@web3-storage/w3infra-ipni",
+  "name": "@web3-storage/w3infra-indexer",
   "version": "0.0.0",
   "type": "module",
   "scripts": {
