+++ indexer/test/lib.block-advert-publisher.spec.js
