+++ indexer/test/lib.block-index-writer.spec.js
