+++ package.json
@@ -12,7 +12,7 @@
     "console": "sst console",
     "lint": "tsc && eslint '**/*.js'",
     "clean": "rm -rf dist node_modules package-lock.json ./*/{.cache,dist,node_modules}",
-    "test": "npm test -w billing -w upload-api -w carpark -w replicator -w roundabout -w filecoin",
+    "test": "npm test -w billing -w upload-api -w carpark -w replicator -w roundabout -w filecoin -w indexer",
     "test-integration": "ava --verbose --serial --timeout=660s --fail-fast test/*.test.js",
     "fetch-metrics-for-space": "npm run fetch-metrics-for-space -w tools",
     "follow-filecoin-receipt-chain": "npm run follow-filecoin-receipt-chain -w tools",
@@ -118,7 +118,7 @@
     "billing",
     "carpark",
     "filecoin",
-    "ipni",
+    "indexer",
     "replicator",
     "roundabout",
     "tools",
